refactored all functions and component related to telegram notifications
