+++ frontend/src/pages/dashboard/Dashboard.jsx
@@ -1,23 +1,18 @@
 import { useState } from "react"
 import HealthIcon from "@/assets/icons/health.svg?react"
 import EthIcon from "@/assets/icons/ethereum.svg?react"
-import TelegramIcon from "@/assets/icons/telegram_dashboard.svg?react"
-import { ActionModal } from "@/components/ui/action-modal"
 import Card from "@/components/ui/card/Card"
 import { Button } from "@/components/ui/custom-button/Button"
 import Spinner from "@/components/ui/spinner/Spinner"
 import { useCheckPosition, useClosePosition } from "@/hooks/useClosePosition"
 import useDashboardData from "@/hooks/useDashboardData"
-import useTelegramNotification from "@/hooks/useTelegramNotification"
 import { useWalletStore } from "@/stores/useWalletStore"
 import DashboardLayout from "../DashboardLayout"
 import DashboardInfoCard from "@/components/dashboard/dashboardCard/DashboardInfoCard"
+import TelegramNotification from "@/components/ui/telegram-notification/TelegramNotification"
 
 export default function DashboardPage({ telegramId }) {
   const { walletId } = useWalletStore()
-  const [showModal, setShowModal] = useState(false)
-  const handleOpen = () => setShowModal(true)
-  const handleClose = () => setShowModal(false)
 
   const { 
     cardData, 
@@ -26,13 +21,17 @@ export default function DashboardPage({ telegramId }) {
     currentSum, 
     depositedData, 
     isLoading } = useDashboardData()
-  const { mutate: closePositionEvent, isLoading: isClosing } = useClosePosition(walletId)
-  const { data: positionData } = useCheckPosition()
-  const { subscribe } = useTelegramNotification()
+
+  const { 
+    mutate:
+     closePositionEvent,
+    isLoading: isClosing } = useClosePosition(walletId)
+
+  const { 
+    data: positionData } = useCheckPosition()
 
   const hasOpenedPosition = positionData?.has_opened_position
 
-  const handleSubscribe = () => subscribe({ telegramId, walletId })
 
   return (
     <DashboardLayout>
@@ -42,12 +41,11 @@ export default function DashboardPage({ telegramId }) {
         <Card label="Borrow Balance" cardData={cardData} icon={<EthIcon className="icon" />} />
       </div>
       <div className="dashboard-info-container">
-
         <DashboardInfoCard
-        cardData={cardData}
-        startSum={startSum}
-        currentSum={currentSum}
-        depositedData={depositedData}
+          cardData={cardData}
+          startSum={startSum}
+          currentSum={currentSum}
+          depositedData={depositedData}
         />
 
         <Button
@@ -57,26 +55,10 @@ export default function DashboardPage({ telegramId }) {
           onClick={() => closePositionEvent()}
           disabled={isClosing || !hasOpenedPosition}
         >
-          {isClosing ? "Closing..." : "Redeem"}
+          {isClosing ?
+           "Closing..." : "Redeem"}
         </Button>
-        <Button variant="secondary" size="lg" className="dashboard-btn telegram" onClick={handleOpen}>
-          <TelegramIcon className="tab-icon" />
-          Enable telegram notification bot
-        </Button>
-        {showModal && (
-          <ActionModal
-            isOpen={showModal}
-            title="Telegram Notification"
-            subTitle="Do you want to enable telegram notification bot?"
-            content={[
-              "This will allow you to receive quick notifications on your telegram line in realtime. You can disable this setting anytime.",
-            ]}
-            cancelLabel="Cancel"
-            submitLabel="Yes, Sure"
-            submitAction={handleSubscribe}
-            cancelAction={handleClose}
-          />
-        )}
+        <TelegramNotification telegramId={telegramId}/>
       </div>
     </DashboardLayout>
   )
